refactor: split codebase into apps/* + packages/* workspaces
Three new packages and one new app:

apps/
  server/      was src/server.ts + test/ + scripts/ + (root) otlp*.js, providers/, pricing.js, logger.js
  dashboard/   was frontend/ (moved in earlier commit)
packages/
  db/          was src/db.ts → packages/db/src/index.ts
  providers/   was providers/*.js → packages/providers/src/*.js
  otlp/        was otlp.js, otlp-logs.js, otlp-metrics.js, otlp-export.js → packages/otlp/src/{traces,logs,metrics,export}.js
  pricing/     was pricing.js + pricing.fallback.json
  shared/      was logger.js → packages/shared/src/logger.js + barrel index.js
  sdk/         (already moved in earlier commit)

Imports updated:
- apps/server/src/server.ts uses @llmflow/{db,otlp/{traces,logs,metrics,export},providers,providers/passthrough,pricing,shared/logger}
- packages/otlp/src/{traces,logs,metrics}.js now require @llmflow/db (was relative './src/db')
- packages/otlp/src/traces.js requires @llmflow/pricing (was './pricing')
- packages/otlp/src/export.js requires @llmflow/shared/logger (was './logger')
- packages/pricing/src/index.js: fallback JSON path is one dir up (now lives in packages/pricing/, not root)
- packages/providers/src/passthrough.js: replace 'uuid' npm dep with built-in crypto.randomUUID()
- apps/server/test/*: stale ../providers, ../otlp-export, ../src/db requires retargeted to workspace paths
- apps/server/test/run-tests.js: pin DASHBOARD_PORT=3000, PROXY_PORT=8080 so the HEALTH_URL poll matches

Each workspace package has its own package.json declaring main + exports +
intra-workspace deps via 'workspace:*'. Root devDeps host typescript +
@playwright/test + bun-types. Per-package node_modules contain the resolved
symlinks (Bun's strict-style hoisting).

The publicDir lookup in server.ts now walks three levels up
(apps/server/src/ → repo root) so the dashboard build artifacts at /public/
keep serving from the same on-disk location.

Verified:
- bunx tsc --noEmit -p apps/server/tsconfig.json: only the 2 pre-existing
  db.ts SQLQueryBindings errors that exist on main.
- bun run apps/server/src/server.ts: boots, prints dashboard + proxy URLs.
- bun run --filter @llmflow/server test: 55 passed (same baseline as main;
  provider e2e suite fails identically, needs real API keys).

Not touched in this commit:
- AGENTS.md, README.md, DOCKER_HUB.md, RELEASING.md still reference old
  paths — follow-up commit will refresh docs.
- providers/ remains CJS — TS port is a separate concern.
