+++ apps/server/scripts/generate-og-image.js
