+++ packages/otlp/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@llmflow/otlp",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./src/traces.js",
+  "exports": {
+    ".": "./src/traces.js",
+    "./traces": "./src/traces.js",
+    "./logs": "./src/logs.js",
+    "./metrics": "./src/metrics.js",
+    "./export": "./src/export.js"
+  },
+  "dependencies": {
+    "@llmflow/db": "workspace:*",
+    "@llmflow/pricing": "workspace:*",
+    "@llmflow/shared": "workspace:*"
+  }
+}
