+++ packages/pricing/src/index.js
@@ -7,7 +7,7 @@ const PRICING_URL = process.env.PRICING_URL || DEFAULT_PRICING_URL;
 const REFRESH_INTERVAL_MS = 24 * 60 * 60 * 1000; // 24 hours
 const VERBOSE = process.env.VERBOSE === '1' || process.argv.includes('--verbose');
 
-const fallbackPricingPath = path.join(__dirname, 'pricing.fallback.json');
+const fallbackPricingPath = path.join(__dirname, '..', 'pricing.fallback.json');
 
 let pricingData = {};
 let lastFetchTime = 0;
