+++ packages/providers/src/azure.js
