+++ packages/providers/src/openai-compatible.js
