+++ packages/providers/src/openai.js
