+++ packages/shared/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@llmflow/shared",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./src/index.js",
+  "exports": {
+    ".": "./src/index.js",
+    "./logger": "./src/logger.js"
+  }
+}
