+++ packages/shared/src/index.js
@@ -1,0 +1,5 @@
+// @llmflow/shared — convenience re-exports.
+// Sub-paths are also exported directly (see package.json#exports).
+module.exports = {
+    log: require('./logger'),
+};
