+++ release-notes/v0.3.2.md
@@ -30,7 +30,8 @@ This patch release fixes critical issues reported during testing, particularly a
 
 **Root Cause:** The API endpoint wasn't extracting the `provider` query param, and `db.getTraces()` had no SQL WHERE clause for provider filtering.
 
-**Fix:** 
+**Fix:**
+
 - Added `provider` and `tag` filter support to `db.getTraces()`
 - Added query param extraction in `/api/traces` and `/api/traces/export`
